# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI Boilerplate",
    "summary": "A modular FastAPI starter with a plugin system",
    "description": "\n    # FastAPI Boilerplate\n\n    A modern FastAPI starter with:\n\n    * Vertical-slice modules and a clean infrastructure layer\n    * Session-based auth with OAuth providers\n    * Swappable cache, queue, and rate-limit backends\n    * SQLAdmin admin UI\n    ",
    "contact": {
      "name": "Benav Labs",
      "url": "https://github.com/benavlabs/FastAPI-boilerplate",
      "email": "contact@benav.io"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.18.0"
  },
  "paths": {
...
    "/api/v1/users/{username}/tier": {
      "get": {
        "tags": [
          "Users",
          "Users"
        ],
        "summary": "Get User Subscription Tier",
        "description": "Retrieves detailed information about a user's subscription tier.\n\n            This endpoint returns comprehensive data about the user's current\n            subscription tier, including name, features, limitations, and any\n            custom configurations.\n\n            Permission rules:\n            - Users can view their own tier information\n            - Administrators can view any user's tier information\n\n            This is useful for displaying subscription information to users\n            or for determining available features in client applications.",
        "operationId": "get_user_tier_api_v1_users__username__tier_get",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tier information retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get User Tier Api V1 Users  Username  Tier Get"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to view this tier information"
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users",
          "Users"
        ],
        "summary": "Update User Subscription Tier (Admin)",
        "description": "Changes a user's subscription tier.\n\n            This admin-only endpoint allows changing which subscription tier\n            a user is assigned to. This affects the user's:\n            - API rate limits\n            - Available features\n            - Access privileges\n\n            When a user's tier is changed, all related configurations (such as\n            rate limits) are automatically updated based on the new tier's settings.",
        "operationId": "update_user_tier_api_v1_users__username__tier_patch",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Username"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "session_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserTierUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User tier updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Update User Tier Api V1 Users  Username  Tier Patch"
                }
              }
            }
          },
          "400": {
            "description": "Invalid tier ID"
          },
          "403": {
            "description": "Not authorized - requires admin privileges"
          },
          "404": {
            "description": "User not found or tier not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tiers/": {
      "get": {
        "tags": [
          "Tiers",
          "Tiers"
        ],
        "summary": "List tiers",
        "description": "Paginated list of tiers.",
        "operationId": "get_tiers_api_v1_tiers__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "items_per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Items Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedData_TierRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tiers/{name}": {
      "get": {
        "tags": [
          "Tiers",
          "Tiers"
        ],
        "summary": "Get a tier by name",
        "description": "Get a tier by name.",
        "operationId": "get_tier_by_name_api_v1_tiers__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TierRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedData_TierRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TierRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedData[TierRead]"
      },
...
      "TierRead": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "Name of the tier",
            "examples": [
              "free",
              "basic",
              "pro",
              "enterprise"
            ]
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at"
        ],
        "title": "TierRead",
        "description": "Schema for reading tier data."
      },
...
      "UserTierUpdate": {
        "properties": {
          "tier_id": {
            "type": "integer",
            "title": "Tier Id"
          }
        },
        "type": "object",
        "required": [
          "tier_id"
        ],
        "title": "UserTierUpdate",
        "description": "Schema for updating a user's tier."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}